# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Aasim API",
        "description": "API documentation for Aasim platform",
        "contact": {
            "email": "support@aasim.com"
        },
        "version": "1.0.0"
    },
    "paths": {
...
        "/hitl-approvals": {
            "get": {
                "tags": [
                    "HITL Approvals"
                ],
                "summary": "List approvals",
                "operationId": "getHitlApprovals",
                "parameters": [
                    {
                        "name": "status",
                        "in": "query",
                        "description": "Filter by status",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    },
                    {
                        "name": "assigned_to",
                        "in": "query",
                        "description": "Filter by assigned user ID",
                        "required": false,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    },
                    {
                        "name": "priority",
                        "in": "query",
                        "description": "Filter by priority",
                        "required": false,
                        "schema": {
                            "type": "string"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "type": "array",
                                            "items": {
                                                "type": "object"
                                            }
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hitl-approvals/{hitlApproval}": {
            "get": {
                "tags": [
                    "HITL Approvals"
                ],
                "summary": "Get approval details",
                "operationId": "getHitlApproval",
                "parameters": [
                    {
                        "name": "hitlApproval",
                        "in": "path",
                        "description": "Approval ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Approval not found"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hitl-approvals/{hitlApproval}/approve": {
            "post": {
                "tags": [
                    "HITL Approvals"
                ],
                "summary": "Approve request",
                "operationId": "approveHitlApproval",
                "parameters": [
                    {
                        "name": "hitlApproval",
                        "in": "path",
                        "description": "Approval ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request approved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hitl-approvals/{hitlApproval}/reject": {
            "post": {
                "tags": [
                    "HITL Approvals"
                ],
                "summary": "Reject request",
                "operationId": "rejectHitlApproval",
                "parameters": [
                    {
                        "name": "hitlApproval",
                        "in": "path",
                        "description": "Approval ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "Request rejected successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/hitl-approvals/{hitlApproval}/escalate": {
            "post": {
                "tags": [
                    "HITL Approvals"
                ],
                "summary": "Escalate request",
                "operationId": "escalateHitlApproval",
                "parameters": [
                    {
                        "name": "hitlApproval",
                        "in": "path",
                        "description": "Approval ID",
                        "required": true,
                        "schema": {
                            "type": "string",
                            "format": "uuid"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Request escalated successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object"
                                }
                            }
                        }
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    },
...
}